requirements.txt
fastapi==0.89.1
openai==0.26.4
Pillow==9.4.0
pydantic==1.10.4
tweepy==4.12.1
python-dotenv==1.0.0